Add group invitations to the invitations section
Split Card to several files
