+++ frontend/src/api/invitations/hooks.ts
@@ -7,7 +7,26 @@ export type Invitation = {
   from_user_id: number
   to_user_id: number
   timestamp: string
-}
+} & (
+  | {
+      type: 0
+    }
+  | {
+      type: 1
+      /** group id */
+      associated_id: number
+    }
+  | {
+      type: 2
+      /** group id */
+      associated_id: number
+    }
+  | {
+      type: 3
+      /** event id */
+      associated_id: number
+    }
+)
 
 /** 0-following,1-group invitation, 2-group join, 3-event */
 export type InvitationType = 0 | 1 | 2 | 3
