+++ frontend/src/components/invitations/FollowingCard.tsx
@@ -1,0 +1,27 @@
+import { FC } from "react"
+import Link from "next/link"
+
+import { Invitation } from "@/api/invitations/hooks"
+import { useUser } from "@/api/users/hooks"
+import Avatar from "@/components/Avatar"
+import { ResponseButtons } from "@/components/invitations/ResponseButtons"
+
+export const FollowingCard: FC<{ invitation: Invitation & { type: 0 } }> = ({ invitation }) => {
+  const { user } = useUser(invitation.from_user_id)
+
+  if (!user) return null
+
+  return (
+    <div className={"flex flex-col"}>
+      <Link href={`/user/${invitation.from_user_id}`} className={"flex gap-2 items-center"}>
+        <span className={"w-12"}>
+          <Avatar user={user} size={10} className={""} />
+        </span>
+        <span>
+          <span className={"text-primary"}>{user.username}</span> wants to follow you
+        </span>
+      </Link>
+      <ResponseButtons invitationId={invitation.id} />
+    </div>
+  )
+}
